requirements.txt
absl-py==2.2.2
antlr4-python3-runtime==4.9.3
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
cython_bbox==0.1.5
deep-sort-realtime==1.3.2
effdet==0.4.1
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
grpcio==1.71.0
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
lap==0.5.12
loguru==0.7.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnx==1.17.0
onnxsim==0.4.36
opencv-python==4.11.0.86
ops==2.20.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==6.30.2
psutil==7.0.0
py-cpuinfo==9.0.0
pybind11==2.13.6
pycocotools==2.0.8
Pygments==2.19.1
pyparsing==3.2.3
python-box==7.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scipy==1.15.2
seaborn==0.13.2
setuptools==78.1.0
six==1.17.0
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
timm==1.0.15
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.107
ultralytics-thop==2.0.14
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
